requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
comtypes==1.1.14
flake8==6.0.0
frozenlist==1.3.3
idna==3.4
mccabe==0.7.0
multidict==6.0.4
openai==0.27.4
Pillow==9.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
pypiwin32==223
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
requests==2.28.2
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.15
yarl==1.9.1
